remove file rather than try and except

+++ executorlib/task_scheduler/file/shared.py
@@ -116,11 +116,9 @@ def execute_tasks_h5(
                     cache_directory=cache_directory
                 ):
                     file_name = os.path.join(cache_directory, task_key + "_i.h5")
-                    try:
-                        dump(file_name=file_name, data_dict=data_dict)
-                    except ValueError:
+                    if os.path.exists(file_name):
                         os.remove(file_name)
-                        dump(file_name=file_name, data_dict=data_dict)
+                    dump(file_name=file_name, data_dict=data_dict)
                     if not disable_dependencies:
                         task_dependent_lst = [
                             process_dict[k] for k in future_wait_key_lst
